Consolidate search hooks into single hook
