+++ context/app/static/js/components/entity-search/Search/hooks.js
@@ -1,9 +1,17 @@
-import { useState, useCallback, useContext } from 'react';
-import { AppContext } from 'js/components/Providers';
-
+import { useState, useCallback, useContext, useMemo } from 'react';
+import { useSearchkitVariables } from '@searchkit/client';
+import { CustomQuery } from '@searchkit/sdk';
 import useDeepCompareEffect from 'use-deep-compare-effect';
 
+import useSearchkitSDK from 'js/components/entity-search/searchkit-modifications/useSearchkitSDK';
+import RequestTransporter from 'js/components/entity-search/searchkit-modifications/RequestTransporter';
+import { AppContext } from 'js/components/Providers';
+import { getAuthHeader } from 'js/helpers/functions';
+import { useStore } from 'js/components/entity-search/SearchWrapper/store';
+import { createSearchkitFacet } from 'js/components/entity-search/SearchWrapper/utils';
 import { fetchSearchData } from 'js/hooks/useSearchData';
+
+import { buildSortPairs, getRangeProps } from './utils';
 
 function useAllResultsUUIDs() {
   const [queryBody, setQueryBody] = useState({});
@@ -35,4 +43,68 @@ function useAllResultsUUIDs() {
   return { allResultsUUIDs, setQueryBodyAndReturnBody };
 }
 
-export { useAllResultsUUIDs };
+const query = new CustomQuery({
+  queryFn: (q) => {
+    return {
+      bool: {
+        must: [
+          {
+            simple_query_string: {
+              fields: ['all_text', 'description'],
+              query: q.match(/^\s*HBM\S+\s*$/i) ? `"${q}"` : q,
+            },
+          },
+        ],
+      },
+    };
+  },
+});
+
+function useSearch(numericFacetsProps) {
+  const { elasticsearchEndpoint, groupsToken } = useContext(AppContext);
+  const authHeader = getAuthHeader(groupsToken);
+  const { fields, facets, defaultFilters, entityType } = useStore();
+
+  const defaultFilterValues = Object.values(defaultFilters);
+
+  const { allResultsUUIDs, setQueryBodyAndReturnBody } = useAllResultsUUIDs();
+
+  const config = useMemo(() => {
+    return {
+      host: elasticsearchEndpoint,
+      connectionOptions: {
+        headers: {
+          ...authHeader,
+        },
+      },
+      hits: {
+        fields: Object.values(fields).map(({ identifier }) => identifier),
+      },
+      sortOptions: buildSortPairs(Object.values(fields)),
+      query,
+      facets: Object.values(facets).map((facet) =>
+        createSearchkitFacet({ ...facet, ...getRangeProps(facet.field, numericFacetsProps) }),
+      ),
+      filters: defaultFilterValues.map((filter) => filter.definition),
+      postProcessRequest: setQueryBodyAndReturnBody,
+    };
+  }, [
+    authHeader,
+    defaultFilterValues,
+    elasticsearchEndpoint,
+    facets,
+    fields,
+    numericFacetsProps,
+    setQueryBodyAndReturnBody,
+  ]);
+
+  const transporter = new RequestTransporter(config);
+
+  const variables = useSearchkitVariables();
+  const defaultSort = 'mapped_last_modified_timestamp.keyword.desc';
+  const { results } = useSearchkitSDK(config, variables, transporter, defaultFilterValues, defaultSort);
+
+  return { results, entityType, allResultsUUIDs };
+}
+
+export { useSearch };
